requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.1.0
langchain-groq==0.0.1
groq==0.2.9
httpx==0.24.1
chromadb==0.4.18
pydantic==2.5.2
pypdf==3.17.1
PyPDF2==3.0.0
python-docx==0.8.11
werkzeug==2.3.7
wkhtmltopdf==0.2
pdfkit==1.0.0
PyMuPDF==1.23.8  # Better PDF handling
sentence-transformers==2.2.2  # Better embeddings
psutil==5.9.0  # For process management
